# swagger/oas30_templ.json
{"openapi":"3.0.0","info":{"title":"srv-logs","version":"1.0.0","description":"Logs Service for AnyPodcast"},"paths":{"/logrec":{"get":{"operationId":"get_all","responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogRecList"}}}}},"parameters":[{"name":"UserID","in":"query","required":true,"schema":{"type":"string"}},{"name":"SearchValue","in":"query","required":false,"schema":{"type":"string"}}],"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"post":{"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogRec"}}},"required":true},"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}}},"/logrec/{id}":{"get":{"operationId":"get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogRec"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"put":{"operationId":"update","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LogRec"}}},"required":true},"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}},"delete":{"operationId":"delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"200 response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/Empty"}}}}},"x-amazon-apigateway-integration":{"httpMethod":"POST","uri":"arn:aws:apigateway:${region}:lambda:path/2015-03-31/functions/arn:aws:lambda:${region}:${account}:function:${lambda_name}/invocations","responses":{"default":{"statusCode":"200"}},"passthroughBehavior":"when_no_match","contentHandling":"CONVERT_TO_TEXT","type":"aws_proxy"}}}},"components":{"schemas":{"Empty":{"title":"Empty Schema","type":"object"},"LogLevel":{"type":"number","enum":[0,1,2,3],"title":"LogLevel"},"LogRec":{"type":"object","properties":{"Created":{"type":"string","default":""},"UserID":{"type":"string","default":""},"Msg":{"type":"string","default":""},"Data":{"type":"string","default":""},"LogType":{"type":"number","default":0}},"required":["Data","LogType","Msg","UserID"],"title":"LogRec"},"LogRecList":{"type":"array","items":{"$ref":"#/components/schemas/LogRec"},"title":"LogRecList"},"SearchParams":{"type":"object","properties":{"UserID":{"type":"string"},"SearchValue":{"type":"string"}},"required":["UserID"],"title":"SearchParams"}}},"tags":[],"servers":[],"x-amazon-apigateway-request-validators":{"Validate query string parameters and headers":{"validateRequestParameters":true,"validateRequestBody":false}}}